Automatically backed up by Learn

+++ lib/create.sql
@@ -1,6 +1,6 @@
 CREATE TABLE projects (
 title TEXT,
 category TEXT,
-funding goal INTEGER,
+funding goal REAL,
 start_date INTEGER,
 end_date INTEGER)
